Refactored names and locations of keybindings components.  Added API documentation.
