+++ TheSadRogue.Integration.Tests/Components/PlayerKeybindingsTest.cs
@@ -1,5 +1,5 @@
 using SadRogue.Primitives;
-using SadRogue.Integration.Components;
+using SadRogue.Integration.Components.Keybindings;
 using Xunit;
 
 namespace SadRogue.Integration.Tests.Components
@@ -7,10 +7,10 @@ namespace SadRogue.Integration.Tests.Components
     public class PlayerControlsTest
     {
         [Fact]
-        public void NewPlayerControlsComponent()
+        public void NewPlayerKeybindingsComponent()
         {
             var player = new RogueLikeEntity((0,0), Color.White,1);
-            var component = new PlayerControlsComponent();
+            var component = new PlayerKeybindingsComponent();
 
             Assert.Equal(4, component.Motions.Count);
             Assert.Empty(component.Actions);
